feat(client@PaperSheet/Show): async select/search term
