+++ client/src/components/PaperSheet/New/TermSelect/Option.js
@@ -11,7 +11,7 @@ const Option = (props): Node => (
     selected={props.isFocused}
     component="div"
     style={{
-      fontWeight: props.isSelected ? 500 : 400,
+      fontWeight: props.isSelected ? 500 : 400
     }}
     {...props.innerProps}
   >
